Create change
Uploaded patch set 1.

Patch-set: 1
Change-id: I2c86fd4fc134afa12de1fee12a52279f1a9fdf31
Subject: Created a client-type enumeration as part of the original COPS refactor and have changed it back to an short value due to a misinterpretation of the COPS specifications. Additionally changed the enumeration held by the COPSDecision to not leverage the ordinal value as a mapping between the COPS payload which comes in as two bytes.
Branch: refs/heads/master
Commit: dfdeb8a916e0c36bdc1708145c158802872f46e3