Update patch set 1
Patch Set 1: Code-Review+1

As these objects will be used as a key to a Map, I need to take all of the bytes of the data and padding separately as equivalent objects may for instance have 6 bytes in data and 2 in padding where the equivalent may contain the same 8 bytes in data and 0 in padding. Therefore, by simply taking the hash of data + padding (which was being done previously) will not take the same hash value.

Patch-set: 1
Reviewer: Gerrit User 3588 <3588@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=+1