Remove "response" from yang.
Once pushing an object to the DS/config, if it's config is not correct,
instead of updating the object by setting the "response" field to know the reason,
the object is deleted and the message that was previously shown in
the "response" is logged at the error level.

This patch also reconciliate CCAP on restart.

Change-Id: I92fbe38d4519d21079341b618e244db8e51cc6c5
Signed-off-by: Alexis de Talhouët <adetalhouet@inocybe.com>