Update patch set 3
Patch Set 3:

(3 comments)

Hi Ryan,

Thanks for the review. I have a few comments related to IGateSpec.java and PacketcableProvider.java. On the latter, I'll just reorganize the logic to eliminate the redundant check. As for the duplicate call, I did that to update the timestamp, there is more detail in my specific replies to your comments.

I need to look at the northbound consumers of that ServiceFlowDirection Yang definition and see if they would be impacted by a change there.

Patch-set: 3
Label: Verified=0