Update patch set 3
Patch Set 3:

okay updating the time stamp makes since to me now.

With respect to the Direction enum. 
In my opinion since the ordinals of the enums do not match using an ordinal to set direction is a source for future errors and GateSpec.setDirection() should take in a Direction not a byte (the constructor takes in the enum). 
Additionally setGate(byte) is never used (direction is only set via constructor with the current code) so there is almost zero impact changing it.

I also feel that the yang file should be changed to match what the cops layer expects (up=0 and down=1). Consumers of the restconf api should not be affected since they should be using the enum names. Some of the other fixes mentioned in the meeting sounded like they would change the yang file so maybe we can get that change included with that update?

Patch-set: 3
Label: Verified=0