Update patch set 2
Patch Set 2:

(1 comment)

The logic seems solid.

There is a historical typo in one log message.

Also, what do you think the best policy is for converting tabs to spaces? I noticed it in the review and hadn't thought to change this type of thing previously. Now that I look, the guidance from ODL is 4 space indentation (which you are enforcing). I assume this is a small issue we should be fixing whenever we find it.

Patch-set: 2
Label: Code-Review=0